backend/requirements.txt
fastapi
pydantic
uvicorn[standard]
sqlalchemy
psycopg2-binary
pwdlib[argon2]
python-dotenv
email-validator
itsdangerous